Fix the time calculation for the update() methods.
That was not precise at all, so the game was updated slower when something
was running slow (connexion or graphics).

+++ src/gui/camera/camera.hpp
@@ -70,7 +70,7 @@ private:
   /**
    * The speed at which the camera moves.
    */
-  uint movement_speed;
+  float movement_speed;
   /**
    * The position of the cursor. It is used to know the movement the camera
    * has to do, when middle click is pressed and the mouse is moved.
